fix(cli): incorrect "set equalizer" value range
-120 to 135 should be allowed, but since the min/max values were not
multiplied by 10, it was -12 to 13.

+++ cli/src/set.rs
@@ -29,7 +29,7 @@ pub async fn set(set_command: SetCommand, device: &impl Device) -> openscq30_lib
             }
         }
         SetCommand::Equalizer { volume_adjustments } => {
-            let adjustment_array: [i8; 8] =
+            let adjustment_array: [i16; 8] =
                 volume_adjustments
                     .try_into()
                     .unwrap_or_else(|values: Vec<_>| {
